pyproject.toml
[project]
name = "dynoai"
version = "1.2.0"
description = "AI-powered dyno tuning toolkit for Harley-Davidson ECM optimization"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "DynoAI Team" }
]
keywords = ["dyno", "tuning", "harley-davidson", "ecm", "ve-table", "motorcycle"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pyyaml>=6.0.0",
    "jsonschema>=4.23.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "werkzeug>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "black>=24.8.0",
    "isort>=5.13.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "bandit>=1.7.0",
    "types-requests",
    "types-PyYAML",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/DynoAI"
Documentation = "https://github.com/yourusername/DynoAI/docs"
Repository = "https://github.com/yourusername/DynoAI"
Issues = "https://github.com/yourusername/DynoAI/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dynoai", "api", "external_scrapers"]

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
  | __pycache__
  | temp_selftest
  | experiments
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
skip = [".venv", "node_modules", "__pycache__", "experiments"]
known_first_party = ["dynoai", "api", "io_contracts"]

[tool.ruff]
line-length = 88
target-version = "py39"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "experiments",
    "temp_selftest",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "experiments/",
    "temp_selftest/",
    "tests/",
    "node_modules/",
    ".venv/",
]

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "numpy.*",
    "matplotlib.*",
    "flask.*",
    "werkzeug.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = ["DynoAI_3", "DynoAI_3-main", "node_modules", ".venv", "experiments", "temp_selftest"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["dynoai", "api"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/experiments/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.bandit]
exclude_dirs = ["tests", "experiments", "temp_selftest"]
skips = ["B101"]  # Skip assert warnings

